src/main/webapp/app/entities/poi/route/poi-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IPoi, Poi } from '../poi.model';
import { PoiService } from '../service/poi.service';

import { PoiRoutingResolveService } from './poi-routing-resolve.service';

describe('Poi routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: PoiRoutingResolveService;
  let service: PoiService;
  let resultPoi: IPoi | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(PoiRoutingResolveService);
    service = TestBed.inject(PoiService);
    resultPoi = undefined;
  });

  describe('resolve', () => {
    it('should return IPoi returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultPoi = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultPoi).toEqual({ id: 123 });
    });

    it('should return new IPoi if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultPoi = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultPoi).toEqual(new Poi());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Poi })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultPoi = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultPoi).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
